+++ src/main/kotlin/app/revanced/patches/youtube/misc/microg/patch/resource/MicroGResourcePatch.kt
@@ -13,6 +13,9 @@ import app.revanced.patches.youtube.misc.microg.annotations.MicroGPatchCompatibi
 import app.revanced.patches.youtube.misc.microg.shared.Constants.BASE_MICROG_PACKAGE_NAME
 import app.revanced.patches.youtube.misc.microg.shared.Constants.REVANCED_PACKAGE_NAME
 import app.revanced.patches.youtube.misc.settings.resource.patch.SettingsResourcePatch
+import app.revanced.patches.youtube.misc.settings.bytecode.patch.SettingsPatch
+import app.revanced.patches.youtube.misc.settings.framework.components.impl.Preference
+import app.revanced.patches.youtube.misc.settings.framework.components.impl.StringResource
 
 @Name("microg-resource-patch")
 @DependsOn([FixLocaleConfigErrorPatch::class, SettingsResourcePatch::class])
@@ -21,25 +24,14 @@ import app.revanced.patches.youtube.misc.settings.resource.patch.SettingsResourc
 @Version("0.0.1")
 class MicroGResourcePatch : ResourcePatch() {
     override fun execute(data: ResourceData): PatchResult {
-        data.xmlEditor["res/xml/settings_fragment.xml"].use {
-            val settingsElementIntent = it.file.createElement("intent")
-            settingsElementIntent.setAttribute("android:targetPackage", "$BASE_MICROG_PACKAGE_NAME.android.gms")
-            settingsElementIntent.setAttribute("android:targetClass", "org.microg.gms.ui.SettingsActivity")
-
-            val settingsElement = it.file.createElement("Preference")
-            settingsElement.setAttribute("android:title", "@string/microg_settings")
-            settingsElement.appendChild(settingsElementIntent)
-
-            it.file.firstChild.appendChild(settingsElement)
-        }
-
-        val settingsFragment = data["res/xml/settings_fragment.xml"]
-        settingsFragment.writeText(
-            settingsFragment.readText().replace(
-                "android:targetPackage=\"com.google.android.youtube",
-                "android:targetPackage=\"$REVANCED_PACKAGE_NAME"
+        SettingsPatch.addPreference(
+            Preference(
+                StringResource("microg_settings", "MicroG Settings"),
+                Preference.Intent("$BASE_MICROG_PACKAGE_NAME.android.gms", "", "org.microg.gms.ui.SettingsActivity"),
+                StringResource("microg_settings_summary", "Settings for MicroG"),
             )
         )
+        SettingsPatch.renameIntentsTargetPackage(REVANCED_PACKAGE_NAME)
 
         val manifest = data["AndroidManifest.xml"]
         manifest.writeText(