+++ src/main/kotlin/app/revanced/patches/youtube/misc/settings/framework/components/impl/InputType.kt
@@ -1,0 +1,6 @@
+package app.revanced.patches.youtube.misc.settings.framework.components.impl
+
+enum class InputType(val type: String) {
+    STRING("text"),
+    NUMBER("number"),
+}